+++ choco/src/main/java/btrplace/solver/choco/runner/disjoint/splitter/ConstraintSplitterMapper.java
@@ -66,7 +66,7 @@ public class ConstraintSplitterMapper {
         mapper.register(new ReadySplitter());
         mapper.register(new RootSplitter());
         mapper.register(new RunningSplitter());
-        mapper.register(new SequentialVMTransitionsSplitter());
+        mapper.register(new SeqSplitter());
         mapper.register(new SleepingSplitter());
         mapper.register(new SplitSplitter());
         mapper.register(new SpreadSplitter());